+++ w4f/signatures/waf/netscaler.py
@@ -2,7 +2,7 @@
 
 VENDOR = {'name': 'netscaler',
  'headers': {'server': 'ns_[a-z]|netscaler',
-             'via': 'ns-cache',
+             'via': r'\bns-cache',
              'cneonction': None,
              'nncoection': None},
  'cookies': ['^ns_af=', '^citrix_ns_id', '^NSC_']}